requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
appnope==0.1.0            # via ipython
backcall==0.1.0           # via ipython
bitstring==3.1.5
click==7.0
decorator==4.4.0          # via ipython, traitlets
ipdb==0.12
ipython-genutils==0.2.0   # via traitlets
ipython==7.4.0
jedi==0.13.3              # via ipython
paho-mqtt==1.4.0
parso==0.4.0              # via jedi
pexpect==4.7.0            # via ipython
pickleshare==0.7.5        # via ipython
prompt-toolkit==2.0.9     # via ipython
ptyprocess==0.6.0         # via pexpect
pygments==2.3.1           # via ipython
six==1.12.0               # via prompt-toolkit, traitlets
terminaltables==3.1.0
traitlets==4.3.2          # via ipython
wcwidth==0.1.7            # via prompt-toolkit

# The following packages are considered to be unsafe in a requirements file:
# setuptools==42.0.1        # via ipdb, ipython
